% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utils.R
\name{cf2treedata}
\alias{cf2treedata}
\title{Convert CF matrix to tree data structure}
\usage{
cf2treedata(cffile)
}
\arguments{
\item{cffile}{Path to CF matrix file}
}
\value{
Tree data object compatible with ggtree
}
\description{
Converts a conflict-free matrix to a tree data structure compatible with ggtree.
}
\examples{
tree_data <- cf2treedata("path/to/CFMatrix.txt")
}
