% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_internal.R
\name{.add_target_clone}
\alias{.add_target_clone}
\title{Add target clone information}
\usage{
.add_target_clone(leaf_df, cf_tableno0_sorted, target_mut)
}
\description{
Add target clone information
}
\keyword{internal}
