% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_internal.R
\name{.extract_ordering}
\alias{.extract_ordering}
\title{Extract ordering from clone structure}
\usage{
.extract_ordering(clone_structure)
}
\description{
Extract ordering from clone structure
}
\keyword{internal}
