% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhyloSOLIDvis-package.R
\name{.onAttach}
\alias{.onAttach}
\title{Check and warn about converTree installation}
\usage{
.onAttach(libname, pkgname)
}
\description{
Check and warn about converTree installation
}
\keyword{internal}
