% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_internal.R
\name{extract_mutation_order}
\alias{extract_mutation_order}
\title{Extract ordering from clone structure}
\usage{
extract_mutation_order(clone_structure)
}
\description{
Extract ordering from clone structure
}
\keyword{internal}
