% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_internal.R
\name{find_zero_barcode}
\alias{find_zero_barcode}
\title{Find zero barcodes (rows with all zeros)}
\usage{
find_zero_barcode(table)
}
\description{
Find zero barcodes (rows with all zeros)
}
\keyword{internal}
