% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_internal.R
\name{find_zero_mutation}
\alias{find_zero_mutation}
\title{Find zero mutations (columns with all zeros)}
\usage{
find_zero_mutation(table)
}
\description{
Find zero mutations (columns with all zeros)
}
\keyword{internal}
