% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_class.R
\name{PhyloData}
\alias{PhyloData}
\title{PhyloData Class for Storing Phylogenetic Data}
\usage{
PhyloData(
  raw_cf = NULL,
  input_table = NULL,
  sorted_cf = NULL,
  tree = NULL,
  tree_data = NULL,
  tree_no0 = NULL,
  zero_barcodes = NULL,
  annotations = NULL,
  clone_data = NULL,
  flipping_data = NULL,
  total_flipping = NULL
)
}
\description{
An S3 class that stores all phylogenetic data after one-time processing.
This avoids repeated file I/O and computation across different visualization functions.
}
\keyword{internal}
