% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_internal.R
\name{process_annotations}
\alias{process_annotations}
\title{Process annotations for plot (backward compatible)}
\usage{
process_annotations(df_all_info)
}
\description{
Process annotations for plot (backward compatible)
}
\keyword{internal}
