% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_internal.R
\name{process_annotations_with_zero}
\alias{process_annotations_with_zero}
\title{Process annotations and generate colors (with zero barcodes filtering)}
\usage{
process_annotations_with_zero(df_all_info, zero_barcodes)
}
\description{
Process annotations and generate colors (with zero barcodes filtering)
}
\keyword{internal}
