% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_internal.R
\name{recursive_clone_partition}
\alias{recursive_clone_partition}
\title{Recursive clone partition}
\usage{
recursive_clone_partition(
  available_items,
  mat,
  parent_members = NULL,
  by_row = TRUE
)
}
\description{
Recursive clone partition
}
\keyword{internal}
