% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_internal.R
\name{remove_zero_barcode}
\alias{remove_zero_barcode}
\title{Remove zero barcodes from CF matrix}
\usage{
remove_zero_barcode(cffile)
}
\description{
Remove zero barcodes from CF matrix
}
\keyword{internal}
