% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat2tree.R
\name{dat2tree}
\alias{dat2tree}
\title{Convert data frame to phylogenetic tree with branch lengths}
\usage{
dat2tree(dat)
}
\arguments{
\item{dat}{Tree data frame with node, parent, and label columns}
}
\value{
Phylogenetic tree object (phylo class)
}
\description{
Converts a tree data frame to a phylo object where branch lengths represent mutation counts.
}
\examples{
tree <- dat2tree(tree_data_frame)
}
