% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhyloSOLIDvis-package.R
\docType{package}
\name{PhyloSOLIDvis}
\alias{PhyloSOLIDvis-package}
\alias{PhyloSOLIDvis}
\title{PhyloSOLIDvis: Visualization Suite for PhyloSOLID Phylogenetic Trees}
\description{
Generate publication-ready circular phylogenetic tree visualizations with
integrated mutation heatmaps, cell annotations, and genotype flipping markers.
}
\author{
\strong{Maintainer}: Qing Yang \email{yangqing@westlake.edu.cn}

Other contributors:
\itemize{
  \item Westlake University [copyright holder]
}

}
\keyword{internal}
