% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_internal.R
\name{process_annotations}
\alias{process_annotations}
\title{Process annotations and generate colors}
\usage{
process_annotations(df_all_info)
}
\description{
Process annotations and generate colors
}
\keyword{internal}
