% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_mutation_matrix.R
\name{sort_mutation_matrix}
\alias{sort_mutation_matrix}
\title{Sort mutation matrix using recursive clone partitioning}
\usage{
sort_mutation_matrix(df)
}
\arguments{
\item{df}{Input data frame with mutations as columns and cells as rows}
}
\value{
Sorted data frame
}
\description{
Implements a recursive algorithm to sort mutations and cells based on
clonal relationships, optimizing the visualization layout.
}
\examples{
sorted_mat <- sort_mutation_matrix(raw_cf_table)
}
